+++ app/helpers/forest_liana/adapter_helper.rb
@@ -5,5 +5,13 @@ module ForestLiana
       quoted_column_name = ActiveRecord::Base.connection.quote_column_name(column_name)
       "#{quoted_table_name}.#{quoted_column_name}"
     end
+
+    def self.cast_boolean(value)
+      if ActiveRecord::Base.connection.adapter_name == 'MySQL'
+        value === 'true' ? 1 : 0;
+      else
+        value
+      end
+    end
   end
 end